Revisions from PR#751
