+++ guide/src/class.md
@@ -44,22 +44,9 @@ unsafe impl pyo3::PyTypeInfo for MyClass {
 
     #[inline]
     fn type_object() -> std::ptr::NonNull<pyo3::ffi::PyTypeObject> {
-        use std::ptr::NonNull;
         use pyo3::type_object::LazyTypeObject;
         static TYPE_OBJECT: LazyTypeObject = LazyTypeObject::new();
-        TYPE_OBJECT.get_or_init(|| {
-                // automatically initialize the class on-demand
-                let gil = pyo3::Python::acquire_gil();
-                let py = gil.python();
-                let boxed = pyo3::pyclass::create_type_object::<Self>(py, Self::MODULE)?;
-                Ok(unsafe { NonNull::new_unchecked(Box::into_raw(boxed)) })
-            })
-            .unwrap_or_else(|e| {
-                let gil = Python::acquire_gil();
-                let py = gil.python();
-                e.print(py);
-                panic!("An error occurred while initializing class {}", Self::NAME)
-            })
+        TYPE_OBJECT.get_pyclass_type::<Self>()
     }
 }
